feat(routing): cleanup, reorg & fix lints
