+++ routing/src/rib/encapsulation.rs
